Refactor: Updates Typography
The type on the site did not match more recent Independent Software materials.

This fix updates all type to Source Sans Pro.

BREAKING CHANGES: Typographic styles and alignments may be broken/inconsistent for some viewport sizes. This update was a first-pass and will be refined. This fix may also make changes to margins and/or padding across the site.