Refactor: Implements v1 of new hero element
The hero element left a lot to be desired both visually and structurally.

This fix implements an entirely new, flexbox-based hero element. It features a streamlined design that is more consistent with the new styles implemented across the rest of the site.

BREAKING CHANGES:

Alignment and style in the hero may be broken/inconsistent between viewports.